Move code to library called MaterialXglTF and create a separate exe using this library.
