+++ .clang-format
@@ -1,0 +1,17 @@
+AllowShortBlocksOnASingleLine: true
+AllowShortLambdasOnASingleLine: false
+AlwaysBreakTemplateDeclarations: No
+BreakBeforeBraces: Allman
+BreakConstructorInitializers: AfterColon
+BreakConstructorInitializersBeforeComma: true
+ColumnLimit: 0
+Cpp11BracedListStyle: false
+IndentCaseLabels: true
+IndentPPDirectives: BeforeHash
+IndentWidth: 4
+LambdaBodyIndentation: OuterScope
+PointerAlignment: Left
+SortIncludes: false
+SpaceAfterCStyleCast: true
+SpaceInEmptyBlock: true
+UseTab: Never
