+++ source/MaterialXglTF/CgltfMaterialLoader.h
