+++ source/MaterialXglTF/CgtfMaterialLoader.cpp
